requirements.txt
foca==0.6.0
gunicorn==20.0.4
setuptools==46.4.0
PyYAML~=5.3.1
typing~=3.7.4.1
Flask~=1.1.2
kubernetes~=17.17.0
pymongo~=3.10.1
Werkzeug~=1.0.1
requests~=2.23.0
connexion~=2.7.0
mongomock~=3.19.0
pytest~=5.4.3
mock~=4.0.3
yq~=2.12.2
kubernetes
gitpython~=3.1.18